requirements.txt
opencv_contrib_python==4.5.3.56
numpy==1.21.0
cvzone==1.3.7
mediapipe==0.8.6
icecream==2.1.1
Pillow==8.3.1
tensorflow==2.5.0
